// client/assets/scripts/CoreDefine.ts
export enum ResType
{
    /**Spine动画 */
    Spine = 0,
    /**位图 */
    Bitmap = 1,
    /**预制体 */
    Prefab = 2,
    /**音效 */
    AudioClip = 3,
    /**文本 */
    Text = 4,
    /**json */
    Json = 5,
}

export enum Music
{
    MusicVolume = "MusicVolume",
    SoundVolume = "SoundVolume",
    MusicMute = "MusicMute",
    SoundMute = "SoundMute",
}

export enum AudioType
{
    Music = 0,
    Sound = 1,
}